Fix iMX6 I2C EDID query with a delay. Add 32MHz display timing. (#71)
Previously the iMX6 was not waiting long enough for a response from the
EDID query. Adding in an actual stall instead of a tight loop resolves
the issue.

Fixing EDID query subsequently broke the waveshare and adafruit screens,
which use a 32MHz pixel clock that was not present in our Pll/Mpll regs
table. Added a new entry for 32MHz that works flawlessly on SABRE and
well on most HummingBoards.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Henry Beberman <hebeberm@microsoft.com>