Update BoundaryDevices/SabreLite to build in Linux environment (#97)
* Remove Audmux block init code.

This is not an appropriate use of MsftFunctionConfig. Audmux
initialization code will be moved into BoardLib init.

* Update ENET DSM to conform to ASL specification

* Remove MsftFunctionConfig macro

MsftFunctionConfig macro is microsoft asl.exe specific.
Additionally it is not a recognized macro by the
ACPI org. Replace it with the equivalent VendorLong
macro.

* Update USB ASL

Update DSM UUID and ARG handling. Fix invalid GPIO device scope for OTG pin

* Remove windows specific GCC build options